requirements.txt
pandas
xlrd
soynlp
torchtext
spacy